// src/app/task/interfaces/Task.interface.ts
// Generated by https://quicktype.io

import { Image } from "./image.interface";

export interface Task {
    id:       number;
    imageUrl: string;
    name:     string;
    status:   TaskStatus;
    image?: Image;
}

export enum TaskStatus {
    Pendiente = 'Pendiente',
    EnProceso = 'En Proceso',
    Completado = 'Completado'
  }