// src/app/auth/interfaces/user.interface.ts
// Generated by https://quicktype.io

import { Task } from "../../task/interfaces/Task.interface";

export interface LoginResponse {
    csrfToken: string;
    token:     string;
    user:      User;
}

export interface User {
    ageRange: string;
    email:    string;
    id:       number;
    phone:    string;
    userName: string;
}

export interface UserRegister {
    ageRange: string;
    email:    string;
    phone:    string;
    userName: string;
    password: string;
}

export interface UserRegisterWithTasks {
    ageRange: string;
    email:    string;
    phone:    string;
    userName: string;
    password: string;
    tasks:    Task[];
}